+++ src/main/java/org/folio/marc/migrations/exceptions/ApiValidationException.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 public class ApiValidationException extends RuntimeException {
 
   private static final String MSG_TEMPLATE = "Unexpected value '%s' in field '%s'";
+  private static final String NOT_ALLOWED_ACTION = "Not allowed %s action for operation with value '%s' in field '%s'";
 
   private final String fieldName;
   private final String fieldValue;
@@ -16,11 +17,25 @@ public class ApiValidationException extends RuntimeException {
     this.fieldValue = fieldValue;
   }
 
+  public ApiValidationException(String action, String fieldName, String fieldValue) {
+    super(NOT_ALLOWED_ACTION.formatted(action, fieldValue, fieldName));
+    this.fieldName = fieldName;
+    this.fieldValue = fieldValue;
+  }
+
   public static ApiValidationException forOperationType(String value) {
     return new ApiValidationException("operationType", value);
   }
 
   public static ApiValidationException forEntityType(String value) {
     return new ApiValidationException("entityType", value);
+  }
+
+  public static ApiValidationException forOperationStatus(String value) {
+    return new ApiValidationException("status", value);
+  }
+
+  public static ApiValidationException notAllowedSaveForOperationStatus(String value) {
+    return new ApiValidationException("data save", "status", value);
   }
 }